requirements.txt
Django==4.0.3
djangorestframework==3.13.1
drf-spectacular==0.21.2
reportlab==3.6.8